backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
httpx
aiohttp
azure-keyvault-secrets
azure-identity
bcrypt
python-jose[cryptography]
